Contributor Dashboard Completed and Components are seperated
