+++ frontend/src/components/Sidebar.css
@@ -1,0 +1,37 @@
+.sidebar {
+  width: 250px;
+  background-color: #121e1e;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+}
+
+.logo {
+  font-size: 1.5rem;
+  margin-bottom: 40px;
+  color: #ffffff;
+}
+
+.sidebar-nav {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  flex-grow: 1;
+}
+
+.sidebar-nav a {
+  color: #a0a0a0;
+  text-decoration: none;
+  padding: 10px 15px;
+  border-radius: 8px;
+  transition: background-color 0.2s;
+}
+
+.sidebar-nav a:hover {
+  background-color: #2a3a3a;
+}
+
+.sidebar-nav a.active {
+  background-color: #254a4a;
+  color: #40e0d0;
+}
